+++ pygenstability/contrib/__init__.py
@@ -1,0 +1,1 @@
+"""Contrib modules."""
